+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `fibers` dependency
 
 ### Removed
+- Remove CLI support for nwayo < 3.3.x (legacy mode)
 - Remove `___nwayo-recommended___` in boilerplate vendors